Remove internals::macros
The remaining macros are only used in `bitcoin`. Move them there.
