+++ bitcoin/src/blockdata/opcodes.rs
@@ -10,7 +10,7 @@
 
 use core::fmt;
 
-use internals::debug_from_display;
+use crate::internal_macros::debug_from_display;
 
 #[cfg(feature = "serde")]
 use crate::prelude::*;
